// OneDrive/Desktop/ERP/client/src/utils/exporters.ts
export type ShopHeader = {
  shopName: string;
  address?: string;
  phone?: string;
};